feat: Attach default arguments to AST
Macro extension is done in two phases, attaching arguments and expanding
macros. Only in the first phase has a call to `parseArgspec`, so it
seems natural to deal with default arguments in the first phase. We
additionally attach default arguments to the AST during attachMacroArgs
and then it will be picked up by existing macro expanders. This breaks
an invariant that printRaw(attachArgs(X)) = printRaw(X).