feat: Support default arguments referencing other arguments
Now expandMacros can correctly expand macros whose default arguments
reference other arguments. A special care was taken in order to follow
the behavior of xparse in case of circular references. In most of the
cases, xparse throws a compilation error, but it works in case like
`O{#2} O{#1}`. From this, we can reasonably guess that xparse treats
default arguments that directly copies another argument's value in a
special way. This behavior was implemented with a simple algorithm, and
it turns out to be consistent with xparse's behavior in cases considered
in unit tests.